Added library files.
Added the .eli component library and .lib pattern library containing
IN-12a, IN-12b, IN-15a and and IN-15b Nixie tube components and the
PL31a-P socket pattern associated with them. Also contains the K155ID1
BCD Nixie driver chip as a component which uses the standard DIP-16
pattern.